requirements.txt
Django-5
django_countries-7.6.1
celery-5.4.0 